[ACH] : correction facture controller

+++ src/main/java/com/pfa/ecommerce/repository/FactureRepository.java
@@ -1,11 +1,9 @@
 package com.pfa.ecommerce.repository;
 
-import java.util.Optional;
-
 import com.pfa.ecommerce.entities.FactureEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactureRepository extends JpaRepository<FactureEntity,Integer>
+public interface FactureRepository extends JpaRepository<FactureEntity,Long>
 {
 	//Optional<FactureEntity> findById(long code);
